Regions 컴포넌트 Categories 컴포넌트 삭제
- buttons 컴포넌트에 합치기
