+++ src/Button.jsx
@@ -1,13 +1,19 @@
 import React from 'react';
 
-export default function Button({ item, onclick, selectedButton }) {
+export default function Button({ buttons, onclick, selectedButton }) {
   return (
     <div>
-      <button type="button" onClick={() => onclick(item)}>
-        {selectedButton.name === item.name
-          ? `${item.name}${'(V)'}`
-          : `${item.name}`}
-      </button>
+      <ul>
+        {buttons.map((button) => (
+          <li key={button.id}>
+            <button type="button" onClick={() => onclick(button)}>
+              {selectedButton.name === button.name
+                ? `${button.name}${'(V)'}`
+                : `${button.name}`}
+            </button>
+          </li>
+        ))}
+      </ul>
     </div>
   );
 }
