+++ src/RegionsContainer.jsx
@@ -1,15 +1,30 @@
-import { useSelector } from 'react-redux';
+import React from 'react';
 
-import Regions from './Regions';
+import { useDispatch, useSelector } from 'react-redux';
+
+import Button from './Button';
+
+import {
+  selecteRegion,
+} from './actions';
 
 export default function RegionsContainer() {
-  const { regions } = useSelector((state) => ({
+  const dispatch = useDispatch();
+
+  const { regions, selectedRegion } = useSelector((state) => ({
     regions: state.regions,
+    selectedRegion: state.selectedRegion,
   }));
 
+  function handleClick(region) {
+    dispatch(selecteRegion(region));
+  }
+
   return (
-    <Regions
-      regions={regions}
+    <Button
+      buttons={regions}
+      onclick={handleClick}
+      selectedButton={selectedRegion}
     />
   );
 }
